requirements.txt
transformers==4.57.3
torch==2.2.1
sentencepiece==0.2.1
datasets==4.4.1

numpy==1.26.4
schedule==1.2.2
python-dotenv==1.0.1

openai==0.28.1

filelock
fsspec
requests
tqdm
pandas
pyarrow
huggingface-hub

httpx==0.27.0
httpcore==0.17.3

python-telegram-bot==21.5